refactor: move peer and url conversion to `p2p/utils` pkg
