+++ p2p/utils/peer_conversion.go
@@ -1,4 +1,4 @@
-package network
+package utils
 
 import (
 	"fmt"
