+++ p2p/utils/url_conversion.go
@@ -1,4 +1,4 @@
-package network
+package utils
 
 import (
 	"crypto/rand"
@@ -112,7 +112,7 @@ func getPeerIP(hostname string) (net.IP, error) {
 		return peerIP, nil
 	}
 
-	// CONSIDER: using a `/dns<4 or 6>/<hostname>` multiaddr instead of resolving here.
+	// CONSIDERATION: using a `/dns<4 or 6>/<hostname>` multiaddr instead of resolving here.
 	// I attempted using `/dns4/.../tcp/...` and go this error:
 	// > failed to listen on any addresses: [can only dial TCP over IPv4 or IPv6]
 	addrs, err := net.LookupHost(hostname)
@@ -120,9 +120,7 @@ func getPeerIP(hostname string) (net.IP, error) {
 		return nil, newResolvePeerIPErr(hostname, err)
 	}
 
-	// CONSIDER: which address(es) should we use when multiple
-	// are provided in a DNS response?
-	// CONSIDER: preferring IPv6 responses when resolving DNS.
+	// CONSIDERATION: preferring IPv6 responses when resolving DNS.
 	// Return first address which is a parsable IP address.
 	var (
 		validIPs    []net.IP
