+++ p2p/utils/url_conversion_test.go
@@ -1,4 +1,4 @@
-package network
+package utils
 
 import (
 	"fmt"
